--- a/execution/executor-types/Cargo.toml
+++ b/execution/executor-types/Cargo.toml
@@ -37,5 +37,5 @@ default = []
 fuzzing = ["aptos-crypto/fuzzing", "aptos-types/fuzzing"]
 
 [[bench]]
-name = "bench1"
+name = "default"
 harness = false