--- a/execution/executor-types/src/lib.rs
+++ b/execution/executor-types/src/lib.rs
@@ -11,9 +11,12 @@ use aptos_crypto::{
 };
 use aptos_scratchpad::{ProofRead, SparseMerkleTree};
 use aptos_types::{
+    account_config::NEW_EPOCH_EVENT_MOVE_TYPE_TAG,
     block_executor::{config::BlockExecutorConfigFromOnchain, partitioner::ExecutableBlock},
     contract_event::ContractEvent,
+    dkg::DKG_START_EVENT_MOVE_TYPE_TAG,
     epoch_state::EpochState,
+    jwks::OBSERVED_JWK_UPDATED_MOVE_TYPE_TAG,
     ledger_info::LedgerInfoWithSignatures,
     proof::{AccumulatorExtensionProof, SparseMerkleProofExt},
     state_store::{state_key::StateKey, state_value::StateValue},
@@ -32,15 +35,12 @@ use std::{
     cmp::max,
     collections::{BTreeSet, HashMap},
     fmt::Debug,
+    ops::Deref,
     sync::{
         atomic::{AtomicBool, Ordering},
         Arc,
     },
 };
-use std::ops::Deref;
-use aptos_types::account_config::NEW_EPOCH_EVENT_MOVE_TYPE_TAG;
-use aptos_types::dkg::DKG_START_EVENT_MOVE_TYPE_TAG;
-use aptos_types::jwks::OBSERVED_JWK_UPDATED_MOVE_TYPE_TAG;
 
 mod error;
 mod executed_chunk;
@@ -531,7 +531,9 @@ impl ProofRead for ProofReader {
 /// Used in both state sync and consensus to filter the txn events that should be subscribable by node components.
 pub fn should_forward_to_subscription_service(event: &ContractEvent) -> bool {
     let type_tag = event.type_tag();
-    type_tag == OBSERVED_JWK_UPDATED_MOVE_TYPE_TAG.deref() || type_tag == DKG_START_EVENT_MOVE_TYPE_TAG.deref() || type_tag == NEW_EPOCH_EVENT_MOVE_TYPE_TAG.deref()
+    type_tag == OBSERVED_JWK_UPDATED_MOVE_TYPE_TAG.deref()
+        || type_tag == DKG_START_EVENT_MOVE_TYPE_TAG.deref()
+        || type_tag == NEW_EPOCH_EVENT_MOVE_TYPE_TAG.deref()
 }
 
 #[cfg(feature = "bench")]