--- a/types/src/dkg/mod.rs
+++ b/types/src/dkg/mod.rs
@@ -13,6 +13,9 @@ use move_core_types::{
 use rand::CryptoRng;
 use serde::{Deserialize, Serialize};
 use std::{collections::BTreeSet, fmt::Debug};
+use std::str::FromStr;
+use once_cell::sync::Lazy;
+use move_core_types::language_storage::TypeTag;
 
 #[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq, CryptoHasher, BCSCryptoHash)]
 pub struct DKGTranscriptMetadata {
@@ -30,6 +33,10 @@ impl MoveStructType for DKGStartEvent {
     const MODULE_NAME: &'static IdentStr = ident_str!("dkg");
     const STRUCT_NAME: &'static IdentStr = ident_str!("DKGStartEvent");
 }
+
+pub static DKG_START_EVENT_MOVE_TYPE_TAG: Lazy<TypeTag> = Lazy::new(||{
+    TypeTag::from_str(DKGStartEvent::struct_tag().to_string().as_str()).unwrap()
+});
 
 /// DKG transcript and its metadata.
 #[derive(Clone, Serialize, Deserialize, Debug, PartialEq, Eq)]