### Description

```
should_forward_to_subscription_service/old                                                                             
                        time:   [593.01 ns 596.51 ns 600.94 ns]

should_forward_to_subscription_service/new                                                                             
                        time:   [75.756 ns 75.995 ns 76.242 ns]
```